Adjust line heights
For the card title, this is now computed as 24px instead of 20px. The prototype currently has 23px.

For the card subtitle, this is now 20px instead of 24px. The prototype currently has 19px.

For h4, it's now 14px which matches the prototype.

For paragraphs it's 20px, whereas the prototype is 18px.

https://www.figma.com/file/RtHSp2qSI9TJfzxjRmLCOj/Covid-Science-Engine---Design-System?node-id=5%3A1