Fix button keyboard navigation issue
Previously this was nesting a button and an anchor tag, so both could
receive focus. This meant that if you tab through the page by keyboard,
you had to press tab twice to get past the button, and when you first
navigated to the button, it doesn't behave like a button (e.g. you can't
press it by pressing space).

This now uses rails "button_to" helper which actually creates a form.

I removed the focus styles on the button because they aren't very visible,
so this is now using the browser default border.