Switch to usernames instead of email for login
There's no point in using emails if we don't validate them.

This changes the login system to use plain usernames. However, I haven't
completely removed the email field yet, because if we ever need to enable any of
the email-related functionality (like password resets or locked account
notifications) then we will have to use this again.